server/requirements.txt
appdirs==1.4.3
docutils==0.13.1
elasticsearch==5.3.0
elasticsearch-dsl==5.2.0
packaging==16.8
pkg-resources==0.0.0
pyparsing==2.2.0
python-dateutil==2.6.0
python-geohash==0.8.5
PyYAML==3.12
six==1.10.0
urllib3==1.21
